4.21: blogilistan laajennus, step9

+++ osa4/blogilista/controllers/blogs.js
@@ -35,6 +35,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid'})
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        response.status(404).json({ error: 'requested blog does not exist' })
+    }
+
+    if (blog.user.toString() !== user.id.toString()) {
+        response.status(401).json({ error: 'cannot delete blogs added by other users' })
+    }
+    
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
